Stabilize initial learner state to avoid redundant JAX compilation
Initialize target critic parameters and step with the online critic shardings, place scalar settings on the replicated mesh, and initialize temperature explicitly as float32.

Validated on one H200 with batch 8, UTD 20 and eight updates: only the first update compiles; checkpoint save and independent restore pass. Cross-run numerical identity was not established.

+++ expo_ft/networks/temperature.py
@@ -9,6 +9,6 @@ class Temperature(nn.Module):
     def __call__(self) -> jnp.ndarray:
         log_temp = self.param(
             "log_temp",
-            init_fn=lambda key: jnp.full((), jnp.log(self.initial_temperature)),
+            init_fn=lambda key: jnp.full((), jnp.log(self.initial_temperature), dtype=jnp.float32),
         )
         return jnp.exp(log_temp)
